+++ index.d.ts
@@ -15,6 +15,7 @@ declare module "native-base" {
 		}
 
 		interface Picker extends ReactNative.PickerProperties {
+			mode?: "dialog" | "dropdown";
 			iosHeader?: string;
 			inlineLabel?: boolean;
 			headerBackButtonText?: string;
@@ -105,6 +106,7 @@ declare module "native-base" {
 			androidStatusBarColor?: string;
 			iosBarStyle?: ReactNative.StatusBarStyle;
 			hasSegment?: boolean;
+			translucent?: boolean;
 		}
 
 		interface Left {
@@ -170,6 +172,8 @@ declare module "native-base" {
 			scrollEnabled?: boolean;
 			style?: ReactNative.ViewStyle | Array<ReactNative.ViewStyle>;
 			contentContainerStyle?: ReactNative.ViewStyle | Array<ReactNative.ViewStyle>;
+			keyboardShouldPersistTaps?: string;
+		        keyboardDismissMode?: string;
 		}
 		/**
          * see Widget Button.js
@@ -681,46 +685,7 @@ declare module "native-base" {
 			color?: string;
 			inverse?: boolean;
 		}
-		/**
-         * vendor react-native-drawer
-         */
-		interface DrawerStyles {
-			drawer?: ReactNative.ViewStyle | Array<ReactNative.ViewStyle>;
-			main?: ReactNative.ViewStyle | Array<ReactNative.ViewStyle>;
-			drawerOverlay?: ReactNative.ViewStyle | Array<ReactNative.ViewStyle>;
-			mainOverlay?: ReactNative.ViewStyle | Array<ReactNative.ViewStyle>;
-		}
-		interface Drawer {
-			acceptDoubleTap?: boolean;
-			acceptPan?: boolean;
-			acceptTap?: boolean;
-			captureGestures?: boolean;
-			children?: any;
-			open?: boolean;
-			closedDrawerOffset?: number;
-			content?: any;
-			deviceScreen?: ReactNative.ScaledSize;
-			disabled?: boolean;
-			initializeOpen?: boolean;
-			negotiatePan?: boolean;
-			onClose?: Function;
-			onCloseStart?: Function;
-			onOpen?: Function;
-			onOpenStart?: Function;
-			openDrawerOffset?: number;
-			openDrawerThreshold?: number;
-			panCloseMask?: number;
-			panOpenMask?: number;
-			panStartCompensation?: boolean;
-			relativeDrag?: boolean;
-			side?: "left" | "right";
-			styles?: DrawerStyles;
-			tapToClose?: boolean;
-			tweenDuration?: number;
-			tweenEasing?: string;
-			tweenHandler?: Function;
-			type?: "overlay" | "static" | "displace";
-		}
+	
 		interface ScrollableTab {
             goToPage?: Function;
             activeTab?: number;
@@ -834,7 +799,36 @@ declare module "native-base" {
 			body?: React.ReactElement<any>;
 			right?: React.ReactElement<any>;
 			style?: ReactNative.ViewStyle;
-        }
+		}
+		
+		interface Accordion {
+			dataArray: Array<any>;
+			headerStyle?: ReactNative.ViewStyle;
+			contentStyle?: ReactNative.ViewStyle;
+			renderHeader?: (item: any) => React.ReactElement<any>;
+			renderContent?: (item: any) => React.ReactElement<any>;
+			icon?: string;
+			expandedIcon?: string;
+			iconStyle?: ReactNative.TextStyle;
+			expandedIconStyle?: ReactNative.TextStyle;
+			style?: ReactNative.ViewStyle;
+		}
+
+		interface DatePicker {
+			defaultDate?: Date;
+			minimumDate?: Date;
+			maximumDate?: Date;
+			locale?: string;
+			placeHolderText?: string;
+			textStyle?: ReactNative.TextStyle;
+			placeHolderTextStyle?: ReactNative.TextStyle;
+			androidMode?: "calendar" | "spinner" | "default";
+			timeZoneOffsetInMinutes?: number;
+			modalTransparent?: boolean;
+			animationType?: "slide" | "fade" | "none";
+			disabled:? boolean;
+			onDateChange?: (date: any) => void;
+		}
 	}
 
 	// Export definitions
@@ -1054,10 +1048,6 @@ declare module "native-base" {
      */
 	export class ProgressBar extends React.Component<NativeBase.ProgressBar, any> {}
 	/**
-     * NativeBase.Drawer
-     */
-	export class Drawer extends React.Component<NativeBase.Drawer, any> {}
-	/**
      * NativeBase.ScrollableTab
      */
 	export class ScrollableTab extends React.Component<NativeBase.ScrollableTab, any> { }
@@ -1120,10 +1110,14 @@ declare module "native-base" {
 			position?: "top" | "bottom" | "center";
 			type?: "danger" | "success" | "warning";
 			duration?: number;
-			onClose?: (reason: "user" | "timeout") => any;;
+			onClose?: (reason: "user" | "timeout") => any;
 			textStyle?: ReactNative.TextStyle;
 			buttonTextStyle?: ReactNative.TextStyle;
 			buttonStyle?: ReactNative.ViewStyle;
 		}): void;
 	}
+	
+	export class Accordion extends React.Component<NativeBase.Accordion, any>{ }
+
+	export class DatePicker extends React.Component<NativeBase.DatePicker, any> { }
 }