+++ src/basic/Item.js
@@ -33,8 +33,6 @@ class Item extends Component {
       if (this.inputProps && this.inputProps.value) {
         this.setState({ isFocused: true });
         this.floatUp(-16);
-      } else {
-        this.setState({ isFocused: false });
       }
       if (this.inputProps && this.inputProps.getRef) {
         this.inputProps.getRef(this._inputRef);
@@ -55,8 +53,6 @@ class Item extends Component {
       if (this.inputProps && this.inputProps.value) {
         this.setState({ isFocused: true });
         this.floatUp(-16);
-      } else {
-        this.setState({ isFocused: false });
       }
       if (this.inputProps && this.inputProps.getRef) {
         this.inputProps.getRef(this._inputRef);
@@ -172,6 +168,69 @@ class Item extends Component {
         return true;
       });
 
+    // if (this.props.floatingLabel && icon.length) {
+    //   let isIcon = false;
+    //   for (let i = 0; i < this.props.children.length; i++) {
+    //     if (this.props.children[i].props.name && this.props.children[i].type.displayName !== "Styled(Input)") {
+    //       isIcon = true;
+    //       newChildren.push(
+    //         <Icon key={[i]} {...this.props.children[i].props} />
+    //       );
+    //     }
+    //     if (this.props.children[i].props.children) {
+    //       newChildren.push(
+    //         <Animated.View
+    //           key="float"
+    //           style={{
+    //             position: "absolute",
+    //             left:
+    //               this.props.last && isIcon
+    //                 ? 40
+    //                 : this.props.last
+    //                   ? 15
+    //                   : isIcon
+    //                     ? 26
+    //                     : 0,
+    //             right: 0,
+    //             top: this.state.topAnim,
+    //             opacity: this.state.opacAnim,
+    //             paddingTop: Platform.OS === "ios" ? undefined : undefined,
+    //             paddingBottom: Platform.OS === "ios" ? undefined : 12
+    //           }}
+    //         >
+    //           <Label {...labelProps}>
+    //             {this.renderLabel(label, labelProps)}
+    //           </Label>
+    //         </Animated.View>
+    //       );
+    //       newChildren.push(
+    //         <Input
+    //           ref={c => (this._inputRef = c)}
+    //           key="l2"
+    //           {...inputProps}
+    //           onFocus={() => {
+    //             this.setState({ isFocused: true });
+    //             inputProps.onFocus && inputProps.onFocus();
+    //           }}
+    //           onBlur={() => {
+    //             inputProps.value
+    //               ? this.setState({
+    //                 isFocused: true
+    //               })
+    //               : !this.state.text.length &&
+    //               this.setState({ isFocused: false });
+    //             inputProps.onBlur && inputProps.onBlur();
+    //           }}
+    //           onChangeText={text => {
+    //             this.setState({ text });
+    //             inputProps.onChangeText && inputProps.onChangeText(text);
+    //           }}
+    //         />
+    //       );
+    //     }
+    //   }
+    // }
+
     if (this.props.floatingLabel) {
       const hasIcon = icon && icon.length;
       if (hasIcon) {
@@ -262,7 +321,7 @@ class Item extends Component {
   }
 }
 
-const childrenType = function(props, propName, component) {
+const childrenType = function (props, propName, component) {
   let error;
   const prop = props[propName];
   if (!props.children.length) {